Struct pca9685_lib::PCA9685

pub struct PCA9685 { /* fields omitted */ }

A representation of a PCA9685 Chip

Methods

impl PCA9685

pub fn new(address: u8, bus: I2c) -> Result<PCA9685, Error>

Creates a new PCA9865

pub async fn start<'_>(&'_ mut self) -> Result<(), Error>

Start the PCA9865. The chip needs a little time to start.

pub fn sleep(&mut self) -> Result<(), Error>

Put the chip into sleep

pub async fn set_prescale_fr<'_>(
&'_ mut self,
frequency: u16,
restart: bool
) -> Result<(), Error>

Set the prescale value from a given frequency.

Warnings

In order to change the prescale, the chip must be put into sleep. Make sure that anything important be safetied before use.

This function tries to be as close as possible to the given frequency, but it is not 100% accurate.

pub fn read_prescale(&mut self) -> Result<u8, Error>

Reads the prescale directly from the chip.

pub fn set_channel(
&mut self,
channel: u8,
period_on: (u8, u8),
period_off: (u8, u8)
) -> Result<(), Error>

Set the pulse-widths for a channel. Channels range from 0 - 15. Since the device uses 12-bit accuracy, the tuples are arranged as (most_significant, least_significant)

Panics

The channel must be less than 16.

pub async fn set_external_clock<'_>(
&'_ mut self,
clock_speed: u32
) -> Result<(), Error>

Set to use External Clock

Warnings

In order to use the EXTCLK pin, the chip must be put to sleep.

In order to reset this mode, you have to run a power cycle (or software reset).

Max frequency is 50 Mhz

Untested

pub fn set_output_mode(&mut self, open_drain: bool) -> Result<(), Error>

Set the output mode of the chip.

Options:

Open-Drain

Totem pole <-- chip default

Warnings

LEDS with built in zener diodes should only be driven in open drain mode.

Untested

pub fn get_prescale(&mut self) -> Result<u8, Error>

Gets the prescale value